:sparkles: 테스트용 stat data 클래스 제거 및 스테미나 관련 코드 수정
제거 사항:
- 실제 사용할 Stat 데이터를 파싱하는 기능을 추가했으므로, Stat에 관한 임시 데이터 클래스를 제거함

추가 사항:
- 스테미나 로직 수정
- stat 데이터를 랩핑하여 다루는 코드 추가

+++ Assets/Dev/Scripts/Item.cs
@@ -11,11 +11,11 @@ public class Item : ScriptableObject
     [SerializeField] private string _name;
     [SerializeField]private Sprite _icon;
 
-    [SerializeField]private float _fillFood; 
-    [SerializeField]private float _fillHealth;
+    [SerializeField]private int _fillFood; 
+    [SerializeField]private int _fillHealth;
 
-    public float FillFood => _fillFood;
-    public float FillHealth => _fillHealth;
+    public int FillFood => _fillFood;
+    public int FillHealth => _fillHealth;
 
     public string Name => _name;
     public Sprite Icon => _icon;
